refactor: 실시간 회의 리팩토링
