+++ src/pages/meeting/InProgressPage.tsx
@@ -1,16 +1,13 @@
-import { useMutation } from "@tanstack/react-query";
-import { useCallback, useEffect, useRef, useState } from "react";
+import { useRef } from "react";
 import { useLocation, useNavigate, useParams } from "react-router-dom";
-import { endMeeting } from "@/apis/meetings";
-import IconClock from "@/assets/icons/media/ic_clock.svg?react";
-import IconCircleUser from "@/assets/icons/user/ic_circle_user.svg?react";
-import { Icon } from "@/components/common/Icon";
 import LiveTranscriptPanel from "./components/LiveTranscriptPanel";
 import MeetingControls from "./components/MeetingControls";
+import MeetingHeader from "./components/MeetingHeader";
 import ParticipantGrid from "./components/ParticipantGrid";
 import { useElapsedTime } from "./hooks/useElapsedTime";
+import { useEndMeeting } from "./hooks/useEndMeeting";
+import { useLocalSpeechFallback } from "./hooks/useLocalSpeechFallback";
 import { useMeetingDetail } from "./hooks/useMeetingDetail";
-import type { InterimEntry, TranscriptEntry } from "./hooks/useMeetingRoom";
 import { useMeetingRoom } from "./hooks/useMeetingRoom";
 import { useSpeechRecognition } from "./hooks/useSpeechRecognition";
 
@@ -21,17 +18,20 @@ const formatStartDateTime = (iso: string) => {
   return `${d.getFullYear()}.${pad(d.getMonth() + 1)}.${pad(d.getDate())} ${pad(d.getHours())}:${pad(d.getMinutes())}`;
 };
 
+const parseMeetingId = (raw: string | undefined): number | null => {
+  if (!raw) return null;
+  const num = Number(raw);
+  return Number.isNaN(num) ? null : num;
+};
+
 const InProgressPage = () => {
   const { meetingId: meetingIdParam } = useParams<{ meetingId: string }>();
   const location = useLocation();
   const navigate = useNavigate();
 
-  const meetingId = (() => {
-    if (!meetingIdParam) return null;
-    const num = Number(meetingIdParam);
-    return Number.isNaN(num) ? null : num;
-  })();
+  const meetingId = parseMeetingId(meetingIdParam);
   const { data: meetingDetail } = useMeetingDetail(meetingId);
+
   const meetingName =
     meetingDetail?.name ??
     (location.state as { name?: string } | null)?.name ??
@@ -40,6 +40,7 @@ const InProgressPage = () => {
     ? formatStartDateTime(meetingDetail.startDateTime)
     : null;
   const memberCount = meetingDetail?.memberCount ?? 0;
+
   const startedAt = useRef(Date.now()).current;
   const elapsed = useElapsedTime(startedAt);
 
@@ -60,88 +61,20 @@ const InProgressPage = () => {
     displayName: "나",
   });
 
-  // 로컬 음성 인식 state (WS 미연결 시 UI에 직접 표시)
-  const [localTranscripts, setLocalTranscripts] = useState<TranscriptEntry[]>(
-    [],
-  );
-  const [localInterim, setLocalInterim] = useState<string>("");
-
-  // WS가 연결되면 로컬 interim 초기화
-  useEffect(() => {
-    if (isWsConnected) {
-      setLocalInterim("");
-    }
-  }, [isWsConnected]);
-
-  // 마이크 음소거 시 로컬 interim 즉시 제거
-  useEffect(() => {
-    if (!isMicEnabled) {
-      setLocalInterim("");
-    }
-  }, [isMicEnabled]);
-
-  const handleInterim = useCallback(
-    (text: string) => {
-      const sent = sendMessage("audio_text", text);
-      if (!sent) {
-        // WS 미연결: 로컬에 직접 표시
-        setLocalInterim(text);
-      }
-    },
-    [sendMessage],
-  );
-
-  const handleFinal = useCallback(
-    (text: string) => {
-      setLocalInterim("");
-      const sent = sendMessage("speech", text);
-      if (!sent) {
-        // WS 미연결: 로컬 transcript에 추가
-        setLocalTranscripts((prev) => [
-          ...prev,
-          {
-            id: `local-${Date.now()}-${prev.length}`,
-            memberId: null,
-            fromName: "나",
-            text,
-            timestamp: new Date().toISOString(),
-            isFinal: true,
-          },
-        ]);
-      }
-    },
-    [sendMessage],
-  );
+  const { handleInterim, handleFinal, mergeTranscripts, mergeInterim } =
+    useLocalSpeechFallback({
+      isWsConnected,
+      isMicEnabled,
+      sendMessage,
+    });
 
   const { isSupported: isSpeechSupported } = useSpeechRecognition({
-    enabled: isMicEnabled,
+    enabled: isMicEnabled, // 마이크 음소거 시 음성 인식도 정지
     onInterim: handleInterim,
     onFinal: handleFinal,
   });
 
-  // 표시용 merged 데이터
-  // - WS 연결됨: WS echo transcript + WS interimByMember
-  // - WS 미연결: 로컬 transcript + 로컬 interim
-  const allTranscripts = [...localTranscripts, ...transcripts];
-
-  const mergedInterimByMember: Record<string, InterimEntry> =
-    localInterim && !isWsConnected
-      ? {
-          ...interimByMember,
-          self: {
-            memberId: null,
-            fromName: "나",
-            text: localInterim,
-            timestamp: new Date().toISOString(),
-          },
-        }
-      : interimByMember;
-
-  const endMutation = useMutation({
-    mutationFn: (id: number) => endMeeting(id),
-    onSuccess: () => navigate("/"),
-    onError: () => navigate("/"),
-  });
+  const { endMeeting } = useEndMeeting();
 
   if (meetingId == null) {
     return null;
@@ -153,50 +86,21 @@ const InProgressPage = () => {
       : [{ id: "self", name: "나", isSelf: true }];
 
   const handleEnd = () => {
-    if (meetingId != null) {
-      endMutation.mutate(meetingId);
-    } else {
-      navigate("/");
-    }
+    if (meetingId != null) endMeeting(meetingId);
+    else navigate("/");
   };
 
   return (
     <div className="flex h-full flex-col py-10">
-      {/* Header */}
-      <div className="flex items-center justify-between">
-        <div className="flex flex-col gap-1">
-          <h1 className="typo-h5 text-(--color-text-primary)">{meetingName}</h1>
-          {meetingStart && (
-            <span className="typo-caption text-(--color-text-tertiary)">
-              {meetingStart}
-              {memberCount > 0 ? ` · 참여자 ${memberCount}명` : ""}
-            </span>
-          )}
-        </div>
-        <div className="flex items-center gap-3">
-          <span className="flex items-center gap-1 rounded-full bg-green-50 px-3 py-1">
-            <Icon icon={IconClock} size={16} className="text-green-700" />
-            <span className="typo-body6 text-green-700">{elapsed}</span>
-          </span>
-          <span className="typo-caption text-(--color-text-tertiary)">
-            WS {isWsConnected ? "✓" : "…"} · RTC {isRoomConnected ? "✓" : "…"}
-          </span>
-          <div className="flex -space-x-2">
-            {displayParticipants.slice(0, 3).map((p) => (
-              <span
-                key={p.id}
-                className="flex h-7 w-7 items-center justify-center rounded-full border-2 border-(--color-bg-surface) bg-light-500"
-              >
-                <Icon
-                  icon={IconCircleUser}
-                  size={20}
-                  className="text-(--color-dark-100)"
-                />
-              </span>
-            ))}
-          </div>
-        </div>
-      </div>
+      <MeetingHeader
+        meetingName={meetingName}
+        meetingStart={meetingStart}
+        memberCount={memberCount}
+        elapsed={elapsed}
+        isWsConnected={isWsConnected}
+        isRoomConnected={isRoomConnected}
+        participants={displayParticipants}
+      />
 
       {errorMessage && (
         <div className="mt-3 rounded-md bg-red-50 px-3 py-2 text-sm text-red-700">
@@ -204,19 +108,17 @@ const InProgressPage = () => {
         </div>
       )}
 
-      {/* Body */}
       <div className="mt-10 flex flex-1 gap-6 overflow-hidden">
         <div className="flex flex-1 flex-col items-center justify-center">
           <ParticipantGrid participants={displayParticipants} />
         </div>
         <LiveTranscriptPanel
-          transcripts={allTranscripts}
-          interimByMember={mergedInterimByMember}
+          transcripts={mergeTranscripts(transcripts)}
+          interimByMember={mergeInterim(interimByMember)}
           isSupported={isSpeechSupported}
         />
       </div>
 
-      {/* Bottom controls */}
       <div className="mt-6 flex justify-center">
         <MeetingControls
           onEnd={handleEnd}
